refactor: extract homepage layout to layouts/homepage.py
