+++ app32.py
@@ -14,6 +14,8 @@ import base64
 
 from utils.performance import calculate_case_performance, calculate_task_performance
 
+from layouts.homepage import build_homepage_layout
+
 
 from db.connection import get_db_connection
 from db.queries import (
@@ -243,55 +245,6 @@ def build_cases_list_component(privilege: str):
         case_items.append(item)
 
     return dbc.ListGroup(case_items, flush=True)
-
-
-def build_homepage_layout(privilege: str):
-    template_types_df = db_fetch_template_types()
-    template_types = template_types_df['type_name'].tolist() if not template_types_df.empty else []
-
-    return html.Div([
-        dcc.Store(id='edit-case-id-store'),
-        dcc.Store(id='delete-case-id-store'),
-        html.Div(id='home-alert-container'),
-
-        dbc.Card(dbc.CardBody([
-            html.H4("Add a New Case", className="card-title"),
-            dbc.Row([
-                dbc.Col(dmc.TextInput(id='home-new-case-name', placeholder="Enter Case Name..."), md=4),
-                dbc.Col(dmc.Select(id='home-new-case-status', data=['Not Started', 'In Progress', 'On Hold'],
-                                   value='Not Started', placeholder="Select Status"), md=3),
-                dbc.Col(dmc.Select(id='home-new-case-type', data=template_types, placeholder="Select Case Type"), md=3),
-                dbc.Col(dmc.Button("Add Case", id='home-add-case-button'), md=2, className="align-self-end"),
-            ])
-        ])),
-        html.H2("All Cases", className="text-center my-4"),
-        html.Div(id='case-list-container', children=build_cases_list_component(privilege)),
-
-        dbc.Modal([
-            dbc.ModalHeader("Edit Case"),
-            dbc.ModalBody([
-                html.Div([dmc.Text("Case Name:", size="sm"), dmc.TextInput(id='modal-edit-case-name')]),
-                html.Div([dmc.Text("Status:", size="sm", mt="sm"), dmc.Select(id='modal-edit-case-status',
-                                                                              data=['Not Started', 'In Progress',
-                                                                                    'On Hold', 'Completed'])]),
-                html.Div([dmc.Text("Case Type:", size="sm", mt="sm"),
-                          dmc.Select(id='modal-edit-case-type', data=template_types)]),
-            ]),
-            dbc.ModalFooter(dmc.Group([
-                dmc.Button("Cancel", id="cancel-edit-case-button", variant="outline"),
-                dmc.Button("Save Changes", id="save-edit-case-button")
-            ]))
-        ], id='edit-case-modal', is_open=False, centered=True),
-
-        dbc.Modal([
-            dbc.ModalHeader("Confirm Deletion"),
-            dbc.ModalBody(id='delete-case-confirm-text'),
-            dbc.ModalFooter(dmc.Group([
-                dmc.Button("Cancel", id='cancel-delete-case-button', variant="outline"),
-                dmc.Button("DELETE", id='confirm-delete-case-button', color='red')
-            ]))
-        ], id='delete-case-modal', is_open=False, centered=True),
-    ])
 
 
 def build_case_detail_layout(case_id: int, username: str, privilege: str):
